Private

Private Documentation

Contents

Private Documentation

Contents

Sphere Functions

Legendre Functions

Healpix

Pixel Covariance

Miscellaneous Utilities

Sphere Functions

CMB.Sphere.x̂ — Constant.

const x̂ = SVector(1, 0, 0)
const ŷ = SVector(0, 1, 0)
const ẑ = SVector(0, 0, 1)

Constant unit vectors in the Cartesian directions.

source

CMB.Sphere.ŷ — Constant.

const x̂ = SVector(1, 0, 0)
const ŷ = SVector(0, 1, 0)
const ẑ = SVector(0, 0, 1)

Constant unit vectors in the Cartesian directions.

source

CMB.Sphere.ẑ — Constant.

const x̂ = SVector(1, 0, 0)
const ŷ = SVector(0, 1, 0)
const ẑ = SVector(0, 0, 1)

Constant unit vectors in the Cartesian directions.

source

CMB.Sphere.:∥ — Method.

∥(u, v)

Test whether vector $u$ is parallel to vector $v$. Assumes that both are unit normalized. See also ⟂.

source

CMB.Sphere.:⟂ — Method.

⟂(u, v)

Test whether vector $u$ is perpendicular to vector $v$. Assumes that both are unit normalized. See also ∥.

source

Legendre Functions

CMB.Legendre.Plm_00 — Function.

Plm_00(::N, ::Type{T}) where {N<:AbstractLegendreNorm, T<:Real}

Returns the initial condition $P_0^0(x)$ for the associated Legendre recursions based on the normalization choice N for numeric type T.

source

CMB.Legendre.Plm_α — Function.

Plm_α(norm::N, ::Type{T}, l::Integer, m::Integer) where {N<:AbstractLegendreNorm, T<:Real}

Returns the coefficient $α_ℓ^m$ for the two-term recursion relation

\[ P_{ℓ+1}^{m}(x) = α_{ℓ+1}^m x P_ℓ^m(x) - β_{ℓ+1}^m P_{ℓ-1}^m(x)\]

where $α_ℓ^m$ is appropriate for the choice of normalization N.

source

CMB.Legendre.Plm_β — Function.

Plm_β(norm::N, ::Type{T}, l::Integer, m::Integer) where {N<:AbstractLegendreNorm, T<:Real}

Returns the coefficient $β_ℓ^m$ for the two-term recursion relation

\[ P_{ℓ+1}^{m}(x) = α_{ℓ+1}^m x P_ℓ^m(x) - β_{ℓ+1}^m P_{ℓ-1}^m(x)\]

where $β_ℓ^m$ is appropriate for the choice of normalization N.

source

CMB.Legendre.Plm_μ — Function.

Plm_μ(norm::N, ::Type{T}, l::Integer) where {N<:AbstractLegendreNorm, T<:Real}

Returns the coefficient $μ_ℓ$ for the single-term recursion relation

\[ P_{ℓ+1}^{ℓ+1}(x) = -μ_{ℓ+1} \sqrt{1-x^2} P_ℓ^ℓ(x)\]

where $μ_ℓ$ is appropriate for the choice of normalization N.

source

CMB.Legendre.Plm_ν — Function.

Plm_ν(norm::N, ::Type{T}, l::Integer) where {N<:AbstractLegendreNorm, T<:Real}

Returns the coefficient $ν_ℓ$ for the single-term recursion relation

\[ P_{ℓ+1}^ℓ(x) = ν_{ℓ+1} x P_ℓ^ℓ(x)\]

where $ν_ℓ$ is appropriate for the choice of normalization N.

source

Healpix

CMB.Healpix.MAX_NSIDE — Constant.

const MAX_NSIDE = 2^29

Maximum valid $N_\mathrm{side}$ parameter value.

source

CMB.Healpix.unsafe_pix2ang — Method.

(θ,ϕ) = unsafe_pix2ang(nside, p)

Like pix2ang but does not call checkhealpix to check nside and pixel index validity.

source

CMB.Healpix.unsafe_pix2phi — Method.

(θ,ϕ) = unsafe_pix2phi(nside, p)

Like pix2phi but does not call checkhealpix to check nside and pixel index validity.

source

CMB.Healpix.unsafe_pix2theta — Method.

(θ,ϕ) = unsafe_pix2theta(nside, p)

Like pix2theta but does not call checkhealpix to check nside and pixel index validity.

source

CMB.Healpix.unsafe_pix2vec — Method.

(θ,ϕ) = unsafe_pix2vec(nside, p)

Like pix2vec but does not call checkhealpix to check nside and pixel index validity.

source

CMB.Healpix.unsafe_pix2z — Method.

(θ,ϕ) = unsafe_pix2z(nside, p)

Like pix2z but does not call checkhealpix to check nside and pixel index validity.

source

Pixel Covariance

CMB.PixelCovariance.FIELDMAP — Constant.

const FIELDMAP

A symbol array which states the canonical ordering of the block matrices within a pixel-pixel covariance matrix.

source

CMB.PixelCovariance.SPECTRAMAP — Constant.

const SPECTRAMAP

A symbol array which states the canonical ordering of spectra.

source

Miscellaneous Utilities

CMB.Util.@abserr — Macro.

@abserr fncall(args...)

Takes the function call fncall(args...) and rewrites the expression to calculate the absolute deviation between the regular call and one with all arguments promoted to BigFloat or BigInt.

Example

julia> CMB.Util.@abserr sin(π-1e-5)
4.1944097844272447e-22

See Also

@relerr, @absrelerr

source

CMB.Util.@absrelerr — Macro.

@absrelerr fncall(args...)

Takes the function call fncall(args...) and rewrites the expression to calculate the absolute and relative deviation between the regular call and one with all arguments promoted to BigFloat or BigInt, returning both as a tuple pair.

Example

julia> CMB.Util.@absrelerr sin(π-1e-5)
(4.1944097844272447e-22, 0.24759425226749643)

See Also

@abserr, @relerr

source

CMB.Util.@relerr — Macro.

@relerr fncall(args...)

Takes the function call fncall(args...) and rewrites the expression to calculate the relative deviation (in ulps) between the regular call and one with all arguments promoted to BigFloat or BigInt.

Example

julia> CMB.Util.@relerr sin(π-1e-5)
0.24759425226749643

See Also

@abserr, @absrelerr

source