Public

Public Documentation

Contents

Public Documentation

Contents

Sphere Functions

Legendre Functions

Healpix

Pixel Covariance

Miscellaneous Utilities

Sphere Functions

CMB.Sphere — Module.

Collection of routines for working with coordinates on the sphere.

source

CMB.Sphere.bearing — Function.

Calculates the bearing angle ($α$), defined as the angle between the meridian (at the first coordinate) and the great circle connecting the first coordinate to the second. Angles are measured clockwise and will be in the range $[0,π)$. See also bearing2.

source

CMB.Sphere.bearing — Method.

α = bearing(θ₁, ϕ₁, θ₂, ϕ₂)

Points on the sphere are given as coordinate pairs $(θ₁,ϕ₁)$ and $(θ₂,ϕ₂)$ where $θ$ is the colatitude angle from the North Pole and $ϕ$ is the azimuthal angle, both in radians.

Examples

julia> bearing(π/2, 0.0, π/4, π/4)
0.6154797086703871

source

CMB.Sphere.bearing — Method.

α = bearing(r₁, r₂)

Points on the sphere are given as unit vectors $r₁$ and $r₂$.

Examples

julia> bearing([1.0, 0.0, 0.0], [0.5, 0.5, sqrt(2)/2])
0.6154797086703873

source

CMB.Sphere.bearing2 — Function.

Calculates the latitude/longitude vector components of the bearing angle (i.e. $δθ = \cos(α), δϕ = \sin(α)$), defined as the angle between the meridian (at the first coordinate) and the great circle connecting the first coordinate to the second. See also bearing.

source

CMB.Sphere.bearing2 — Method.

(δθ, δϕ) = bearing2(θ₁, ϕ₁, θ₂, ϕ₂)

Points on the sphere are given as coordinate pairs $(θ₁,ϕ₁)$ and $(θ₂,ϕ₂)$ where $θ$ is the colatitude angle from the North Pole and $ϕ$ is the azimuthal angle, both in radians.

Examples

julia> bearing2(π/2, 0.0, π/4, π/4)
(0.8164965809277261, 0.5773502691896256)

source

CMB.Sphere.bearing2 — Method.

(δθ, δϕ) = bearing2(r₁, r₂)

Points on the sphere are given as unit vectors $r₁$ and $r₂$.

Examples

julia> bearing2([1.0, 0.0, 0.0], [0.5, 0.5, sqrt(2)/2])
(0.816496580927726, 0.5773502691896257)

source

CMB.Sphere.cosdistance — Function.

Calculates the cosine of the inner angle ($z$) between unit vectors pointing from the center of the sphere to two points on its surface.

source

CMB.Sphere.cosdistance — Method.

z = cosdistance(θ₁, ϕ₁, θ₂, ϕ₂)

Points on the sphere are given as coordinate pairs $(θ₁,ϕ₁)$ and $(θ₂,ϕ₂)$ where $θ$ is the colatitude angle from the North Pole and $ϕ$ is the azimuthal angle, both in radians.

Examples

julia> cosdistance(π/2, 0.0, π/4, π/4)
0.5

source

CMB.Sphere.cosdistance — Method.

z = cosdistance(r₁, r₂)

Points on the sphere are given as unit vectors $r₁$ and $r₂$.

Examples

julia> cosdistance([1.,0.,0.], [0.5,0.5,sqrt(2)/2])
0.5

source

CMB.Sphere.distance — Function.

Calculates the inner angle ($σ$) between unit vectors pointing from the center of the sphere to two points on its surface.

source

CMB.Sphere.distance — Method.

σ = distance(θ₁, ϕ₁, θ₂, ϕ₂)

Points on the sphere are given as coordinate pairs $(θ₁,ϕ₁)$ and $(θ₂,ϕ₂)$ where $θ$ is the colatitude angle from the North Pole and $ϕ$ is the azimuthal angle, both in radians.

Examples

julia> distance(π/2, 0.0, π/4, π/4)
1.0471975511965979

source

CMB.Sphere.distance — Method.

σ = distance(r₁, r₂)

Points on the sphere are given as unit vectors $r₁$ and $r₂$.

Examples

julia> distance([1.,0.,0.], [0.5,0.5,sqrt(2)/2])
1.0471975511965979

source

Legendre Functions

CMB.Legendre — Module.

Collections of functions which compute the associated Legendre functions.

Based on implementation described in Limpanuparb and Milthorpe (2014) “Associated Legendre Polynomials and Spherical Harmonics Computation for Chemistry Applications” arXiv:1410.1748v1

source

CMB.Legendre.AbstractLegendreNorm — Type.

abstract type AbstractLegendreNorm end

Abstract supertype for normalization conditions of the Associated Legendre polynomials.

Example

julia> subtypes(AbstractLegendreNorm)
3-element Array{Union{DataType, UnionAll},1}:
 CMB.Legendre.LegendreNormCoeff
 CMB.Legendre.LegendreSphereNorm
 CMB.Legendre.LegendreUnitNorm

source

CMB.Legendre.LegendreNormCoeff — Type.

struct LegendreNormCoeff{N<:AbstractLegendreNorm,T<:Real} <: AbstractLegendreNorm

Precomputed recursion relation coefficients for the normalization N and value type T.

Example

julia> LegendreNormCoeff{LegendreSphereNorm,Float64}(1)
CMB.Legendre.LegendreNormCoeff{CMB.Legendre.LegendreSphereNorm,Float64} for lmax = 1, mmax = 1 with coefficients:
    μ: [0.0, 1.22474]
    ν: [1.73205, 2.23607]
    α: [0.0 0.0; 1.73205 0.0]
    β: [0.0 0.0; -0.0 0.0]

source

CMB.Legendre.LegendreSphereCoeff — Type.

LegendreSphereCoeff{T}

Table type of precomputed recursion relation coefficients for the spherical harmonic normalization. Alias for LegendreNormCoeff{LegendreSphereNorm,T}.

source

CMB.Legendre.LegendreSphereNorm — Type.

struct LegendreSphereNorm <: AbstractLegendreNorm end

source

CMB.Legendre.LegendreUnitCoeff — Type.

LegendreUnitCoeff{T}

Precomputed recursion relation coefficients for the standard unit normalization. Alias for LegendreNormCoeff{LegendreUnitNorm,T}.

source

CMB.Legendre.LegendreUnitNorm — Type.

struct LegendreUnitNorm <: AbstractLegendreNorm end

source

CMB.Legendre.Nlm — Method.

N = Nlm([T=Float64], l, m)

Computes the normalization constant

\[ N_ℓ^m ≡ \sqrt{\frac{2ℓ+1}{4π} \frac{(ℓ-m)!}{(ℓ+m)!}}\]

which defines the Spherical Harmonic normalized functions $λ_ℓ^m(x)$ in terms of the standard unit normalized $P_ℓ^m(x)$

\[ λ_ℓ^m(x) ≡ N_ℓ^m P_ℓ^m(x)\]

using numbers of type T.

See also Plm and λlm.

source

CMB.Legendre.Pl! — Method.

Pl!(P::AbstractVector, lmax::Integer, x)

Fills the vector P with the Legendre polynomial values $P_ℓ(x)$ for all degrees 0 ≤ ℓ ≤ lmax at x.

source

CMB.Legendre.Pl — Method.

p = Pl(l::Integer, x)

Computes the scalar value $p = P_ℓ(x)$, where $P_ℓ(x)$ is the Legendre polynomial of degree l at x.

source

CMB.Legendre.Plm! — Method.

Plm!(P::AbstractVector, lmax::Integer, m::Integer, x)

Fills the vector P with the Legendre polynomial values $P_ℓ^m(x)$ for all degrees 0 ≤ ℓ ≤ lmax and constant order m at x.

source

CMB.Legendre.Plm! — Method.

Plm!(P::AbstractMatrix, lmax::Integer, mmax::Integer, x)

Fills the lower triangle of the matrix P with the associated Legendre polynomial values $P_ℓ^m(x)$ for all degrees 0 ≤ ℓ ≤ lmax and all orders 0 ≤ m ≤ ℓ at x.

source

CMB.Legendre.Plm — Method.

p = Plm(l::Integer, m::Integer, x)

Computes the scalar value $p = P_ℓ^m(x)$, where $P_ℓ^m(x)$ is the associated Legendre polynomial of degree l and order m at x.

source

CMB.Legendre.legendre! — Method.

legendre!(norm::AbstractLegendreNorm, P::AbstractMatrix, lmax::Integer, mmax::Integer, x)

Fills the matrix Λ with the pre-normalized Legendre polynomial values $N_ℓ^m P_ℓ^m(x)$ for all degrees 0 ≤ ℓ ≤ lmax and all orders 0 ≤ m ≤ ℓ at x, where $N_ℓ^m$ is the normalization scheme norm.

source

CMB.Legendre.legendre! — Method.

legendre!(norm::AbstractLegendreNorm, Λ::AbstractVector, lmax::Integer, m::Integer, x)

Fills the vector Λ with the pre-normalized Legendre polynomial values $N_ℓ^m P_ℓ^m(x)$ for all degrees 0 ≤ ℓ ≤ lmax and constant order m at x, where $N_ℓ^m$ is the normalization scheme norm.

source

CMB.Legendre.legendre — Method.

p = legendre(norm::AbstractLegendreNorm, l::Integer, x)

Computes the scalar value $p = N_ℓ P_ℓ(x)$, where $P_ℓ(x)$ is the Legendre polynomial of degree l at x and $N_ℓ$ is the normalization scheme norm.

source

CMB.Legendre.legendre — Method.

p = legendre(norm::AbstractLegendreNorm, l::Integer, m::Integer, x)

Computes the scalar value $p = N_ℓ^m P_ℓ^m(x)$, where $P_ℓ^m(x)$ is the associated Legendre polynomial of degree l and order m at x and $N_ℓ^m$ the normalization scheme norm.

source

CMB.Legendre.λlm! — Method.

λlm!(Λ::AbstractVector, lmax::Integer, m::Integer, x)

Fills the vector Λ with the spherical harmonic normalized associated Legendre polynomial values $λ_ℓ^m(x)$ for all degrees 0 ≤ ℓ ≤ lmax and constant order m at x.

source

CMB.Legendre.λlm! — Method.

λlm!(Λ::AbstractMatrix, lmax::Integer, mmax::Integer, x::Real)

Fills the lower triangle of the matrix Λ with the spherical harmonic normalized associated Legendre polynomial values $Λ_ℓ^m(x)$ for all degrees 0 ≤ ℓ ≤ lmax and all orders 0 ≤ m ≤ ℓ at x.

source

CMB.Legendre.λlm — Method.

λ = λlm(l::Integer, m::Integer, x)

Computes the scalar value $λ = λ_ℓ^m(x)$, where $λ_ℓ^m(x)$ is the spherical-harmonic normalized associated Legendre polynomial of degree l and order m at x.

source

Healpix

CMB.Healpix — Module.

A module of functions implementing function which interact with the HEALPix pixel definitions. In most cases, only the RING ordering functions are being provided.

See "HEALPix: A Framework for High-Resolution Discretization and Fast Analysis of Data Distributed on the Sphere" Górski, Hivon, & Banday et al (2005) ApJ 622:759–771 arXiv: astro-ph/0409513

source

CMB.Healpix.UNSEEN — Constant.

const UNSEEN = -1.6375e+30

Special value recognized by the libhealpix/healpy routines as an unobserved/masked pixel.

source

CMB.Healpix.InvalidNside — Type.

InvalidNside(nside)

An invalid nside value was provided.

source

CMB.Healpix.InvalidPixel — Type.

InvalidPixel(nside, pix)

An invalid pixel index pix was provided for the given nside.

source

CMB.Healpix.checkhealpix — Method.

checkhealpix(nside, pix)

Throws an InvalidNside exception if nside is not a valid value or an InvalidPixel exception if pix is out of range for the given $N_\mathrm{side}$.

source

CMB.Healpix.checkhealpix — Method.

checkhealpix(nside)

Throws an InvalidNside exception if nside is not a valid value.

source

CMB.Healpix.iscap — Function.

iscap(nside, pix)

Test for whether the given pixel pix is in either polar cap for an nside HEALPix map.

source

CMB.Healpix.isequbelt — Function.

isequbelt(nside, pix)

Test for whether the given pixel pix is in the equatorial belt for an nside HEALPix map.

source

CMB.Healpix.ishealpixok — Method.

isheapixok(nside, pix)

Returns true if nside is valid and pix is in the range 0 to nside2npix(nside) - 1, otherwise false.

source

CMB.Healpix.ishealpixok — Method.

ishealpixok(nside)

Returns true if nside is a power of two in the range 1 to 2^29, otherwise false.

source

CMB.Healpix.isnorth — Function.

isnorth(nside, pix)

Test for whether the given pixel pix is in the northern hemisphere (including the equatorial ring) for an nside HEALPix map.

source

CMB.Healpix.isnorthcap — Function.

isnorthcap(nside, pix)

Test for whether the given pixel pix is in the northern polar cap for an nside HEALPix map.

source

CMB.Healpix.isnorthequbelt — Function.

isnorthequbelt(nside, pix)

Test for whether the given pixel pix is in the northern equatorial belt (including the equatorial ring) for an nside HEALPix map.

source

CMB.Healpix.issouth — Function.

issouth(nside, pix)

Test for whether the given pixel pix is in the southern hemisphere (excluding the equatorial ring) for an nside HEALPix map.

source

CMB.Healpix.issouthcap — Function.

issouthcap(nside, pix)

Test for whether the given pixel pix is in the southern polar cap for an nside HEALPix map.

source

CMB.Healpix.issouthequbelt — Function.

issouthequbelt(nside, pix)

Test for whether the given pixel pix is in the southern equatorial belt (excluding the equatorial ring) for an nside HEALPix map.

source

CMB.Healpix.npix2nside — Method.

Nisde = npix2nside(npix)

Returns the equivalent Nside corresponding to the number of pixels npix.

source

CMB.Healpix.nring2nside — Method.

Nside = nring2nside(nring)

Returns the equivalent Nside corresponding to the number of iso-latitude rings nring.

source

CMB.Healpix.nside2npix — Function.

Npix = nside2npix(nside)

Returns the total number of pixels Npix in an nside HEALPix map. Note that HEALPix pixel indexing is 0-based, so valid pixel values are in the range 0 to Npix - 1.

source

CMB.Healpix.nside2npixcap — Function.

Npix = nside2npixcap(nside)

Returns the number of pixels Npix in the polar caps for the given nside HEALPix map.

source

CMB.Healpix.nside2npixequ — Function.

Npix = nside2npixequ(nside)

Returns the number of pixels Npix in the northern hemisphere, including the equatorial ring, for the given nside HEALPix map.

source

CMB.Healpix.nside2nring — Function.

Nring = nside2nring(nside)

Returns the number of iso-latitude rings Nring in the nside HEALPix map.

source

CMB.Healpix.nside2pixarea — Function.

σ = nside2pixarea(nside)

Returns the surface area σ (in steradians) of each pixel in the given nside HEALPix map.

source

CMB.Healpix.pix2ang — Method.

(θ,ϕ) = pix2ang(nside, p)

Computes the colatitude and azimuth pair (θ,ϕ) for the given pixel p.

source

CMB.Healpix.pix2phi — Method.

ϕ = pix2phi(nside, p)

Computes the azimuth ϕ for the given pixel p. nside is the Nside resolution factor.

source

CMB.Healpix.pix2ring — Method.

i = pix2ring(nside, p)

Computes the ring index i for the given pixel p. nside is the Nside resolution factor.

source

CMB.Healpix.pix2ringidx — Method.

j = pix2ringidx(nside, p)

Computes the index j within the ring for the given pixel p. nside is the Nside resolution factor.

source

CMB.Healpix.pix2theta — Method.

θ = pix2theta(nside, p)

Computes the colatitude θ for the given pixel p. nside is the Nside resolution factor.

source

CMB.Healpix.pix2vec — Method.

r::SVector{3} = pix2vec(nside, p)

Computes the unit vector r pointing to the pixel center of the given pixel p.

source

CMB.Healpix.pix2z — Method.

z = pix2z(nside, p)

Computes the cosine of the colatitude z for the given pixel p. nside is the Nside resolution factor.

source

Pixel Covariance

CMB.PixelCovariance — Module.

Collection of functions which compute the pixel-pixel covariance of the CMB sky.

Based on equations given in Tegmark and de Oliveira-Costa (2001) “How to measure CMB polarization power spectra without losing information” arXiv:astro-ph/0012120v3

source

CMB.PixelCovariance.PixelCovarianceCache — Type.

struct PixelCovarianceCache

Data structure which contains all the information and buffers required to compute the pixel-pixel covariance terms for a given pixel with respect to all other pixels.

source

CMB.PixelCovariance.PixelCovarianceCoeff — Type.

struct PixelCovarianceCoeff{T<:Real}

Precomputed recursion relation coefficients for computing the pixel-pixel covariance.

Example

julia> PixelCovarianceCoeff{Float64}(2)
CMB.PixelCovariance.PixelCovarianceCoeff{Float64} for lmax = 2 with coefficients:
    λ: CMB.Legendre.LegendreNormCoeff{CMB.Legendre.LegendreUnitNorm,Float64}
    η: [0.0795775, 0.238732, 0.397887]
    α: [0.0, 0.0, 0.324874]
    β: [0.0, 0.0, 0.0331573]

source

Miscellaneous Utilities

CMB.Util — Module.

Miscellaneous utility functions.

source

CMB.Util.outer — Function.

Computes the outer product between a given column of a sparse matrix and a vector.

source

CMB.Util.outer — Method.

outer(w::AbstractVector, A::SparseMatrixCSC, n::Integer)

Performs the equivalent of $\vec w \vec{a}_n^\dagger$ where $\vec a_n$ is the column A[:,n].

source

CMB.Util.outer — Method.

outer(A::SparseMatrixCSC, n::Integer, w::AbstractVector)

Performs the equivalent of $\vec a_n \vec w^\dagger$ where $\vec a_n$ is the column A[:,n].

source

CMB.Util.quadprod — Function.

quadprod(A, b, n, dir=:col)

Computes the quadratic product $ABA^T$ efficiently for the case where $B$ is all zero except for the nth column or row vector b, for dir = :col or dir = :row, respectively.

source