Private Documentation

Contents

Private Documentation

Contents

Generic utilities

Sphere Functions

Healpix

Spherical Harmonics

Pixel Covariance

Generic utilities

CMB.quadprod — Function

quadprod(A, b, n, dir=:col)

Computes the quadratic product $ABA^\top$ efficiently for the case where $B$ is all zero except for the nth column or row vector b, for dir = :col or dir = :row, respectively.

Sphere Functions

CMB.Sphere.x̂ — Constant

const x̂ = SVector(1, 0, 0)
const ŷ = SVector(0, 1, 0)
const ẑ = SVector(0, 0, 1)

Constant unit vectors in the Cartesian directions.

CMB.Sphere.ŷ — Constant

const x̂ = SVector(1, 0, 0)
const ŷ = SVector(0, 1, 0)
const ẑ = SVector(0, 0, 1)

Constant unit vectors in the Cartesian directions.

CMB.Sphere.ẑ — Constant

const x̂ = SVector(1, 0, 0)
const ŷ = SVector(0, 1, 0)
const ẑ = SVector(0, 0, 1)

Constant unit vectors in the Cartesian directions.

CMB.Sphere.:∥ — Method

∥(u, v) -> Bool

Test whether vector $u$ is parallel to vector $v$. Assumes that both are unit normalized.

CMB.Sphere.cartvec — Method

r = cartvec(θ, ϕ)
r = cartvec((θ, ϕ))

Converts the colatitude-azimuth pair (θ, ϕ) to a Cartesian unit vector r.

CMB.Sphere.colataz — Method

θ, ϕ = colataz(δ, λ)
θ, ϕ = colataz((δ, λ))

Converts the latitude-longitude pair (δ, λ) in degrees to colatitude-azimuth (θ, ϕ) in radians.

CMB.Sphere.colataz — Method

θ, ϕ = colataz(r)

Converts the Cartesian unit vector r to a colatitude-azimuth pair (θ, ϕ) in radians.

CMB.Sphere.latlon — Method

δ, λ = latlon(θ, ϕ)
δ, λ = latlon((θ, ϕ))

Converts the colatitude-azimuth pair (θ, ϕ) in radians to latitude-longitude (δ, λ) in radians.

CMB.Sphere.latlon — Method

δ, λ = latlon(r)

Converts the Cartesian unit vector r to a latitude-longitude pair (δ, λ) in degrees.

Healpix

CMB.Healpix.MAX_NSIDE — Constant

const MAX_NSIDE = 2^29

Maximum valid $N_\mathrm{side}$ parameter value.

CMB.Healpix.unsafe_ang2pix — Method

p = unsafe_ang2pix(nside, θ, ϕ)

Like ang2pix but neither calls checkhealpix to check the validity of nside nor checks the domain of the spherical coordinates θ and ϕ.

CMB.Healpix.unsafe_pix2ang — Method

(θ,ϕ) = unsafe_pix2ang(nside, p)

Like pix2ang but does not call checkhealpix to check nside and pixel index validity.

CMB.Healpix.unsafe_pix2phi — Method

(θ,ϕ) = unsafe_pix2phi(nside, p)

Like pix2phi but does not call checkhealpix to check nside and pixel index validity.

CMB.Healpix.unsafe_pix2theta — Method

(θ,ϕ) = unsafe_pix2theta(nside, p)

Like pix2theta but does not call checkhealpix to check nside and pixel index validity.

CMB.Healpix.unsafe_pix2vec — Method

(θ,ϕ) = unsafe_pix2vec(nside, p)

Like pix2vec but does not call checkhealpix to check nside and pixel index validity.

CMB.Healpix.unsafe_pix2z — Method

(θ,ϕ) = unsafe_pix2z(nside, p)

Like pix2z but does not call checkhealpix to check nside and pixel index validity.

CMB.Healpix.unsafe_vec2pix — Method

p = unsafe_vec2pix(nside, r)

Like vec2pix but does not check the validity of the nside or length of r.

CMB.Healpix.unsafe_zphi2pix — Method

p = unsafe_zphi2pix(nside, z, ϕ)

Like unsafe_ang2pix but uses the value $z = \cos(θ)$ instead.

Spherical Harmonics

Pixel Covariance