Private Documentation

This page documents private functions which may be useful but which have no guarantee of stability across any change in versions of this package.

Contents

Private Documentation

Contents

Generic utilities

Sphere Functions

Pixel Covariance

Generic utilities

CMB.quadprod — Method

quadprod(A::AbstractSparseMatrixCSC, b::AbstractVecOrMat, n::Integer)

Computes the quadratic product $ABA^\top$ efficiently for the case where $B$ is all zero except for a small number of columns b starting at the nth.

Sphere Functions

CMB.Sphere.x̂ — Constant

const x̂ = SVector(1, 0, 0)
const ŷ = SVector(0, 1, 0)
const ẑ = SVector(0, 0, 1)

Constant unit vectors in the Cartesian directions.

CMB.Sphere.ŷ — Constant

const x̂ = SVector(1, 0, 0)
const ŷ = SVector(0, 1, 0)
const ẑ = SVector(0, 0, 1)

Constant unit vectors in the Cartesian directions.

CMB.Sphere.ẑ — Constant

const x̂ = SVector(1, 0, 0)
const ŷ = SVector(0, 1, 0)
const ẑ = SVector(0, 0, 1)

Constant unit vectors in the Cartesian directions.

CMB.Sphere.:∥ — Method

∥(u, v) -> Bool

Test whether vector $u$ is parallel to vector $v$. Assumes that both are unit normalized.

CMB.Sphere.cartvec — Method

r = cartvec(θ, ϕ)
r = cartvec((θ, ϕ))

Converts the colatitude-azimuth pair $(θ, ϕ)$ to a Cartesian unit vector $r$.

CMB.Sphere.colataz — Method

θ, ϕ = colataz(δ, λ)
θ, ϕ = colataz((δ, λ))

Converts the latitude-longitude pair $(δ, λ)$ in degrees to colatitude-azimuth $(θ, ϕ)$ in radians.

CMB.Sphere.colataz — Method

θ, ϕ = colataz(r)

Converts the Cartesian unit vector $r$ to a colatitude-azimuth pair $(θ, ϕ)$ in radians.

CMB.Sphere.latlon — Method

δ, λ = latlon(θ, ϕ)
δ, λ = latlon((θ, ϕ))

Converts the colatitude-azimuth pair $(θ, ϕ)$ in radians to latitude-longitude $(δ, λ)$ in radians.

CMB.Sphere.latlon — Method

δ, λ = latlon(r)

Converts the Cartesian unit vector $r$ to a latitude-longitude pair $(δ, λ)$ in degrees.

Pixel Covariance