Public Documentation

Contents

Public Documentation

Contents

Sphere Functions

Healpix

File I/O

Pixel Covariance

Sphere Functions

CMB.Sphere — Module

Collection of routines for working with coordinates on the sphere.

CMB.Sphere.bearing — Function

Calculates the bearing angle ($α$), defined as the angle between the meridian (at the first coordinate) and the great circle connecting the first coordinate to the second. Angles are measured eastward of north and will be in the range $[-π,π]$. See also bearing2.

CMB.Sphere.bearing — Method

α = bearing(θ₁, ϕ₁, θ₂, ϕ₂)

Points on the sphere are given as coordinate pairs $(θ₁,ϕ₁)$ and $(θ₂,ϕ₂)$ where $θ$ is the colatitude angle from the North Pole and $ϕ$ is the azimuthal angle, both in radians.

Examples

julia> bearing(π/2, 0.0, π/4, π/4)
0.6154797086703873

CMB.Sphere.bearing — Method

α = bearing(r₁, r₂)

Points on the sphere are given as unit vectors $r₁$ and $r₂$.

Examples

julia> bearing([1.0, 0.0, 0.0], [0.5, 0.5, sqrt(2)/2])
0.6154797086703873

CMB.Sphere.bearing2 — Function

Calculates the north/east vector components of the bearing angle (i.e. $δn = \cos(α), δe = \sin(α)$), defined as the angle between the meridian (at the first coordinate) and the great circle connecting the first coordinate to the second. See also bearing.

CMB.Sphere.bearing2 — Method

(δn, δe) = bearing2(θ₁, ϕ₁, θ₂, ϕ₂)

Points on the sphere are given as coordinate pairs $(θ₁,ϕ₁)$ and $(θ₂,ϕ₂)$ where $θ$ is the colatitude angle from the North Pole and $ϕ$ is the azimuthal angle, both in radians.

Examples

julia> bearing2(π/2, 0.0, π/4, π/4)
(0.8164965809277261, 0.5773502691896257)

CMB.Sphere.bearing2 — Method

(δn, δe) = bearing2(r₁, r₂)

Points on the sphere are given as unit vectors $r₁$ and $r₂$.

Examples

julia> bearing2([1.0, 0.0, 0.0], [0.5, 0.5, sqrt(2)/2])
(0.816496580927726, 0.5773502691896257)

CMB.Sphere.cosdistance — Function

Calculates the cosine of the inner angle ($z$) between unit vectors pointing from the center of the sphere to two points on its surface.

CMB.Sphere.cosdistance — Method

z = cosdistance(θ₁, ϕ₁, θ₂, ϕ₂)

Points on the sphere are given as coordinate pairs $(θ₁,ϕ₁)$ and $(θ₂,ϕ₂)$ where $θ$ is the colatitude angle from the North Pole and $ϕ$ is the azimuthal angle, both in radians.

Examples

julia> cosdistance(π/2, 0.0, π/4, π/4)
0.5

CMB.Sphere.cosdistance — Method

z = cosdistance(r₁, r₂)

Points on the sphere are given as unit vectors $r₁$ and $r₂$.

Examples

julia> cosdistance([1.,0.,0.], [0.5,0.5,sqrt(2)/2])
0.5

CMB.Sphere.distance — Function

Calculates the inner angle ($σ$) between unit vectors pointing from the center of the sphere to two points on its surface.

CMB.Sphere.distance — Method

σ = distance(θ₁, ϕ₁, θ₂, ϕ₂)

Points on the sphere are given as coordinate pairs $(θ₁,ϕ₁)$ and $(θ₂,ϕ₂)$ where $θ$ is the colatitude angle from the North Pole and $ϕ$ is the azimuthal angle, both in radians.

Examples

julia> distance(π/2, 0.0, π/4, π/4)
1.0471975511965979

CMB.Sphere.distance — Method

σ = distance(r₁, r₂)

Points on the sphere are given as unit vectors $r₁$ and $r₂$.

Examples

julia> distance([1.,0.,0.], [0.5,0.5,sqrt(2)/2])
1.0471975511965979

CMB.Sphere.reckon — Function

Calculates a position on the sphere a given distance (σ, in radians) and relative bearing angle (α, in radians) away from a given point (measuring the eastward-of-north orientation of the great circle connecting the source and destination points with respect to the merdian passing through the source).

CMB.Sphere.reckon — Method

(θ′, ϕ′) = reckon(θ, ϕ, σ, α)

The point on the sphere is given by the colatitude-azimuth pair (θ, ϕ), both given in radians.

Note

When r points to either the north or south pole, the meridian is defined to be θ meridian, and the bearing angle α is oriented with respect to it.

For example, moving a distance $π/2$ with no bearing goes to the equator, with the longitude dependent on the input longitude:

julia> reckon(0.0, 0.0, π/2, 0.0)
(1.5707963267948966, 3.141592653589793)

julia> reckon(0.0, π/2, π/2, 0.0)
(1.5707963267948966, 4.71238898038469)

CMB.Sphere.reckon — Method

r′ = reckon(r::AbstractVector, σ, α)

The point on the sphere is given as a unit vector r.

Note

When r points to either the north or south pole, the meridian is defined to be prime meridian and the bearing angle α is oriented with respect to it.

For example, moving a distance $π/2$ with no bearing goes to the negative $x$ axis (i.e. 0° N, 180° W):

julia> reckon([0.0, 0.0, 1.0], π/2, 0.0)
3-element StaticArrays.SArray{Tuple{3},Float64,1,3} with indices SOneTo(3):
 -1.0
  0.0
  6.123233995736766e-17

Healpix

CMB.Healpix — Module

A module of functions implementing function which interact with the HEALPix pixel definitions. In most cases, only the RING ordering functions are being provided.

See "HEALPix: A Framework for High-Resolution Discretization and Fast Analysis of Data Distributed on the Sphere" Górski, Hivon, & Banday et al (2005) ApJ 622:759–771 arXiv: astro-ph/0409513

CMB.Healpix.UNSEEN — Constant

const UNSEEN = -1.6375e+30

Special value recognized by the libhealpix/healpy routines as an unobserved/masked pixel.

CMB.Healpix.InvalidNside — Type

InvalidNside(nside)

An invalid nside value was provided.

CMB.Healpix.InvalidPixel — Type

InvalidPixel(nside, pix)

An invalid pixel index pix was provided for the given nside.

CMB.Healpix.ang2pix — Method

p = ang2pix(nside, θ, ϕ)

Computes the HEALPix pixel index p which contains the point $(θ,ϕ)$ given by the colatitude θ and azimuth ϕ, where nside is the Nside resolution factor.

CMB.Healpix.checkhealpix — Method

checkhealpix(nside, pix)

Throws an InvalidNside exception if nside is not a valid value or an InvalidPixel exception if pix is out of range for the given $N_\mathrm{side}$.

CMB.Healpix.checkhealpix — Method

checkhealpix(nside)

Throws an InvalidNside exception if nside is not a valid value.

CMB.Healpix.iscap — Function

iscap(nside, pix)

Test for whether the given pixel pix is in either polar cap for an nside HEALPix map.

CMB.Healpix.isequbelt — Function

isequbelt(nside, pix)

Test for whether the given pixel pix is in the equatorial belt for an nside HEALPix map.

CMB.Healpix.ishealpixok — Method

isheapixok(nside, pix)

Returns true if nside is valid and pix is in the range 0 to nside2npix(nside) - 1, otherwise false.

CMB.Healpix.ishealpixok — Method

ishealpixok(nside)

Returns true if nside is a power of two in the range 1 to 2^29, otherwise false.

CMB.Healpix.isnorth — Function

isnorth(nside, pix)

Test for whether the given pixel pix is in the northern hemisphere (including the equatorial ring) for an nside HEALPix map.

CMB.Healpix.isnorthcap — Function

isnorthcap(nside, pix)

Test for whether the given pixel pix is in the northern polar cap for an nside HEALPix map.

CMB.Healpix.isnorthequbelt — Function

isnorthequbelt(nside, pix)

Test for whether the given pixel pix is in the northern equatorial belt (including the equatorial ring) for an nside HEALPix map.

CMB.Healpix.issouth — Function

issouth(nside, pix)

Test for whether the given pixel pix is in the southern hemisphere (excluding the equatorial ring) for an nside HEALPix map.

CMB.Healpix.issouthcap — Function

issouthcap(nside, pix)

Test for whether the given pixel pix is in the southern polar cap for an nside HEALPix map.

CMB.Healpix.issouthequbelt — Function

issouthequbelt(nside, pix)

Test for whether the given pixel pix is in the southern equatorial belt (excluding the equatorial ring) for an nside HEALPix map.

CMB.Healpix.npix2nside — Method

Nisde = npix2nside(npix)

Returns the equivalent Nside corresponding to the number of pixels npix.

CMB.Healpix.nring2nside — Method

Nside = nring2nside(nring)

Returns the equivalent Nside corresponding to the number of iso-latitude rings nring.

CMB.Healpix.nside2npix — Function

Npix = nside2npix(nside)

Returns the total number of pixels Npix in an nside HEALPix map. Note that HEALPix pixel indexing is 0-based, so valid pixel values are in the range 0 to Npix - 1.

CMB.Healpix.nside2npixcap — Function

Npix = nside2npixcap(nside)

Returns the number of pixels Npix in the polar caps for the given nside HEALPix map.

CMB.Healpix.nside2npixequ — Function

Npix = nside2npixequ(nside)

Returns the number of pixels Npix in the northern hemisphere, including the equatorial ring, for the given nside HEALPix map.

CMB.Healpix.nside2nring — Function

Nring = nside2nring(nside)

Returns the number of iso-latitude rings Nring in the nside HEALPix map.

CMB.Healpix.nside2pixarea — Function

σ = nside2pixarea(nside)

Returns the surface area σ (in steradians) of each pixel in the given nside HEALPix map.

CMB.Healpix.pix2ang — Method

(θ,ϕ) = pix2ang(nside, p)

Computes the colatitude and azimuth pair (θ,ϕ) for the given pixel p, where nside is the Nside resolution factor.

CMB.Healpix.pix2phi — Method

ϕ = pix2phi(nside, p)

Computes the azimuth ϕ for the given pixel p, where nside is the Nside resolution factor.

CMB.Healpix.pix2ring — Method

i = pix2ring(nside, p)

Computes the ring index i for the given pixel p, where nside is the Nside resolution factor.

CMB.Healpix.pix2ringidx — Method

j = pix2ringidx(nside, p)

Computes the index j within the ring for the given pixel p, where nside is the Nside resolution factor.

CMB.Healpix.pix2theta — Method

θ = pix2theta(nside, p)

Computes the colatitude θ for the given pixel p, where nside is the Nside resolution factor.

CMB.Healpix.pix2vec — Method

r = pix2vec(nside, p)

Computes the unit vector r pointing to the pixel center of the given pixel p, where nside is the Nside resolution factor.

CMB.Healpix.pix2z — Method

z = pix2z(nside, p)

Computes the cosine of the colatitude z for the given pixel p, where nside is the Nside resolution factor.

CMB.Healpix.vec2pix — Method

p = vec2pix(nside, r)

Computes the HEALPix pixel index p which contains the point at the end of the unit vector r, where nside is the Nside resolution factor.

File I/O

CMB.Files.read_obsmat — Function

read_obsmat(filename::String; keywords...)

Read a sparse observing matrix and the corresponding pixelization descriptors from filename. The return value is a NamedTuple with the following fields:

R: the sparse observing matrix $R$.

pixr: either missing or a description of the "right-hand side" pixelization format of the observing matrix — i.e. the pixelization of a map vector $v$ for which the matrix-vector multiplication $R v$ is defined.

pixl: either missing or a description of the "left-hand side" pixelication format of the observing matrix — i.e. the pixelization of a map vector $w$ which results from the matrix-vector multiplication $w = R v$.

The pixel descriptions pixr and pixl are loaded from the datasets named by the keyword arguments of the same name, if available. If the named dataset exists, then it is loaded and returned. (The data type is format- and situation-specific.) It is not an error for the named dataset to not exist, and the field will be returned with the value missing.

See also: write_obsmat

Extended help

Keywords

The following keywords identify the specific data structures or fields within the given data file that correspond to the observing matrix and pixelization data structures or fields, and they are available on all backends. Additional keywords may be supported on a backend-specific basis.

name: Name of the observing matrix $R$. Defaults to "R".

pixr: Name of the data structure or field describing the right-hand side pixelization format. Defaults to "pixels_right".

pixl: Name of the data structure or field describing the "left-hand side" pixelization format. Defaults to "pixels_right".

In all cases, the field names may be nothing to indicate the corresponding data should not be loaded, and the corresponding field in the returned named tuple will have the value missing. It is an error for name to point to a non-existent field, whereas the pixel descriptions will silently ignore an invalid name and just return missing instead.

Backends

The HDF5.jl storage backend is always loaded with CMB.jl and is considered the native storage format. See write_obsmat for writing a native HDF5 file to disk.

Importing observing matrices from the following additional data formats is supported via Requires.jl, which requires the user to first load the extra backend of choice.

JLD and JLD2-flavored HDF5 files with JLD.jl and JLD2.jl, respectively.

MATLAB v5, v6, v7, and v7.3 save files with MAT.jl.

scipy.sparse CSC and CSR matrices saved to HDF5 files with h5sparse. This case is supported without needing to load any extra packages.

The pixelization descriptions are imported as data format specific types. For instance, all formats support loading strings and simple numerical scalars or arrays. Additionally, JLD and JLD2 formats can load named datasets as arbitrary Julia types, MATLAB structs are deserialized as Dicts, and named HDF5 groups are read as (nested) Dicts.

read_obsmat(file::FileIO.File{format"HDF5"};
            mmap::Union{Bool,Val{true},Val{false}} = READ_OBSMAT_MMAP[],
            keywords...)

Reads an observing matrix and pixelization descriptors from an HDF5 file. The native format as written by write_obsmat as well as h5sparse-formatted data structures are supported.

With HDF5 files in the native format, the additional keyword mmap argument controls whether the observing matrix arrays are loaded via memory-mapping or not. If memory-mapped on Unix systems, the MADV_WILLNEED advice will be applied to the arrays. Additionally, on Linux systems, the memory mapping includes the MAP_POPULATE flag to pre-fault the data pages into active RAM. The default behavior is controlled by the global Ref READ_OBSMAT_MMAP.

read_obsmat(file::FileIO.File{format"JLD"}; keywords...)

Reads an observing matrix and pixelization descriptors from a JLD-formatted HDF5 file. This function is conditionally included via Requires.jl and requires the user to import JLD or using JLD first.

read_obsmat(file::FileIO.File{format"JLD2"}; keywords...)

Reads an observing matrix and pixelization descriptors from a JLD2-formatted HDF5 file. This function is conditionally included via Requires.jl and requires the user to import JLD2 or using JLD2 first.

read_obsmat(file::FileIO.File{format"MAT"}; keywords...)

Reads an observing matrix and pixelization descriptors from a MATLAB save file in v5, v6, v7, or v7.3 format. This function is conditionally included via Requires.jl and requires the user to import MAT or using MAT first.

CMB.Files.READ_OBSMAT_MMAP — Constant

READ_OBSMAT_MMAP = Ref{Bool}(true)

Controls the default memory mapping behavior of read_obsmat. Defaults to true.

CMB.Files.READ_OBSMAT_MMAP_FLAGS — Constant

READ_OBSMAT_MMAP_FLAGS = Ref{UnixMmap.MmapFlags}(MAP_SHARED)

Controls the flags passed to mmap when read_obsmat memory maps the observing matrix arrays. Defaults to MAP_SHARED on all systems, with Linux also including MAP_POPULATE by default.

CMB.Files.write_obsmat — Function

write_obsmat(filename::String, obsmat::SparseMatrixCSC, pixr = missing, pixl = missing;
             obsmat_name::String = "R", pixr_name::String = "pixels_right",
             pixl_name::String = "pixels_left")

Write an observing matrix obsmat to the HDF5 file filename. If the file exists, it will be overwritten completely (to ensure proper data alignment required for memory mapping), otherwise it will be created. Pixelization descriptions pixr and pixl may also be provided; for a missing value, the description will not be written to disk.

The HDF5 dataset and group names of the observing matrix and pixelization descriptions can be changed from their defaults by setting the obsmat_name, pixr_name, and pixl_name keywords which effect the obsmat, pixr, and pixl storage locations, respectively.

See also: read_obsmat

Pixel Covariance

The following enums/bit flags are not exported globally, but all of the named values can be imported into a scope by using the parent module (e.g. to access all of the covariance field constants, use using CMB.PixelCovariance.CovarianceFields.

CMB.PixelCovariance.PolarizationConventions.Convention — Type

@enum Convention IAUConv HealpixConv

An enumeration to specify the two types of polarization conventions used to describe Stokes Q/U coordinate systems.

CMB.PixelCovariance.CovarianceFields.Field — Type

@bitflag Field TT QT UT TQ QQ UQ TU QU UU NO_FIELD=0

A bitfield for identifying subblocks of the pixel-pixel covariance matrix. There are 9 subblocks, named as the Cartesian product of elements T, Q, and U:

TT  TQ  TU
QT  QQ  QU
UT  UQ  UU

CMB.PixelCovariance.CovarianceFields.TPol — Constant

const TPol = QT | UT | TQ | TU

An alias for the temperature-cross-polarization sub-blocks of the full covariance matrix.

CMB.PixelCovariance.CovarianceFields.Pol — Constant

const Pol  = QQ | UQ | QU | UU

An alias for the polarization-only sub-blocks of the full covariance matrix.