Public Documentation

Contents

Public Documentation

Contents

Sphere Functions

Legendre Functions

Healpix

Pixel Covariance

Miscellaneous Utilities

Sphere Functions

CMB.Sphere — Module

Collection of routines for working with coordinates on the sphere.

CMB.Sphere.bearing — Function

Calculates the bearing angle ($α$), defined as the angle between the meridian (at the first coordinate) and the great circle connecting the first coordinate to the second. Angles are measured clockwise and will be in the range $[0,π)$. See also bearing2.

CMB.Sphere.bearing — Method

α = bearing(θ₁, ϕ₁, θ₂, ϕ₂)

Points on the sphere are given as coordinate pairs $(θ₁,ϕ₁)$ and $(θ₂,ϕ₂)$ where $θ$ is the colatitude angle from the North Pole and $ϕ$ is the azimuthal angle, both in radians.

Examples

julia> bearing(π/2, 0.0, π/4, π/4)
0.6154797086703873

CMB.Sphere.bearing — Method

α = bearing(r₁, r₂)

Points on the sphere are given as unit vectors $r₁$ and $r₂$.

Examples

julia> bearing([1.0, 0.0, 0.0], [0.5, 0.5, sqrt(2)/2])
0.6154797086703873

CMB.Sphere.bearing2 — Function

Calculates the latitude/longitude vector components of the bearing angle (i.e. $δθ = \cos(α), δϕ = \sin(α)$), defined as the angle between the meridian (at the first coordinate) and the great circle connecting the first coordinate to the second. See also bearing.

CMB.Sphere.bearing2 — Method

(δθ, δϕ) = bearing2(θ₁, ϕ₁, θ₂, ϕ₂)

Points on the sphere are given as coordinate pairs $(θ₁,ϕ₁)$ and $(θ₂,ϕ₂)$ where $θ$ is the colatitude angle from the North Pole and $ϕ$ is the azimuthal angle, both in radians.

Examples

julia> bearing2(π/2, 0.0, π/4, π/4)
(0.8164965809277261, 0.5773502691896257)

CMB.Sphere.bearing2 — Method

(δθ, δϕ) = bearing2(r₁, r₂)

Points on the sphere are given as unit vectors $r₁$ and $r₂$.

Examples

julia> bearing2([1.0, 0.0, 0.0], [0.5, 0.5, sqrt(2)/2])
(0.816496580927726, 0.5773502691896257)

CMB.Sphere.cosdistance — Function

Calculates the cosine of the inner angle ($z$) between unit vectors pointing from the center of the sphere to two points on its surface.

CMB.Sphere.cosdistance — Method

z = cosdistance(θ₁, ϕ₁, θ₂, ϕ₂)

Points on the sphere are given as coordinate pairs $(θ₁,ϕ₁)$ and $(θ₂,ϕ₂)$ where $θ$ is the colatitude angle from the North Pole and $ϕ$ is the azimuthal angle, both in radians.

Examples

julia> cosdistance(π/2, 0.0, π/4, π/4)
0.5

CMB.Sphere.cosdistance — Method

z = cosdistance(r₁, r₂)

Points on the sphere are given as unit vectors $r₁$ and $r₂$.

Examples

julia> cosdistance([1.,0.,0.], [0.5,0.5,sqrt(2)/2])
0.5

CMB.Sphere.distance — Function

Calculates the inner angle ($σ$) between unit vectors pointing from the center of the sphere to two points on its surface.

CMB.Sphere.distance — Method

σ = distance(θ₁, ϕ₁, θ₂, ϕ₂)

Points on the sphere are given as coordinate pairs $(θ₁,ϕ₁)$ and $(θ₂,ϕ₂)$ where $θ$ is the colatitude angle from the North Pole and $ϕ$ is the azimuthal angle, both in radians.

Examples

julia> distance(π/2, 0.0, π/4, π/4)
1.0471975511965979

CMB.Sphere.distance — Method

σ = distance(r₁, r₂)

Points on the sphere are given as unit vectors $r₁$ and $r₂$.

Examples

julia> distance([1.,0.,0.], [0.5,0.5,sqrt(2)/2])
1.0471975511965979

Legendre Functions

CMB.Legendre — Module

Collections of functions which compute the associated Legendre functions.

Based on implementation described in Limpanuparb and Milthorpe (2014) “Associated Legendre Polynomials and Spherical Harmonics Computation for Chemistry Applications” arXiv:1410.1748v1

CMB.Legendre.AbstractLegendreNorm — Type

abstract type AbstractLegendreNorm end

Abstract supertype for normalization conditions of the Associated Legendre polynomials.

Example

julia> using InteractiveUtils; subtypes(AbstractLegendreNorm)
3-element Array{Any,1}:
 LegendreNormCoeff
 LegendreSphereNorm
 LegendreUnitNorm

CMB.Legendre.LegendreNormCoeff — Type

struct LegendreNormCoeff{N<:AbstractLegendreNorm,T<:Real} <: AbstractLegendreNorm

Precomputed recursion relation coefficients for the normalization N and value type T.

Example

julia> LegendreNormCoeff{LegendreSphereNorm,Float64}(1)
LegendreNormCoeff{LegendreSphereNorm,Float64} for lmax = 1, mmax = 1 with coefficients:
    μ: [0.0, 1.22474]
    ν: [1.73205, 2.23607]
    α: [0.0 0.0; 1.73205 0.0]
    β: [0.0 0.0; -0.0 0.0]

CMB.Legendre.LegendreSphereCoeff — Type

LegendreSphereCoeff{T}

Table type of precomputed recursion relation coefficients for the spherical harmonic normalization. Alias for LegendreNormCoeff{LegendreSphereNorm,T}.

CMB.Legendre.LegendreSphereNorm — Type

struct LegendreSphereNorm <: AbstractLegendreNorm end

Trait type denoting the spherical-harmonic normalization of the associated Legendre polynomials.

CMB.Legendre.LegendreUnitCoeff — Type

LegendreUnitCoeff{T}

Precomputed recursion relation coefficients for the standard unit normalization. Alias for LegendreNormCoeff{LegendreUnitNorm,T}.

CMB.Legendre.LegendreUnitNorm — Type

struct LegendreUnitNorm <: AbstractLegendreNorm end

Trait type denoting the unit normalization of the associated Legendre polynomials.

CMB.Legendre.Nlm — Method

N = Nlm([T=Float64], l, m)

Computes the normalization constant

\[ N_ℓ^m ≡ \sqrt{\frac{2ℓ+1}{4π} \frac{(ℓ-m)!}{(ℓ+m)!}}\]

which defines the Spherical Harmonic normalized functions $λ_ℓ^m(x)$ in terms of the standard unit normalized $P_ℓ^m(x)$

\[ λ_ℓ^m(x) ≡ N_ℓ^m P_ℓ^m(x)\]

using numbers of type T.

See also Plm and λlm.

CMB.Legendre.Pl! — Method

Pl!(P, l::Integer, x)

Fills the array P with the unit-normalized Legendre polynomial values $P_ℓ(x)$ for fixed order $m = 0$; equivalent to legendre!(LegendreUnitNorm(), P, l, 0, x).

CMB.Legendre.Pl — Method

p = Pl(l::Integer, x::Number)

Computes the Legendre polynomials using unit normalization and for degree $m = 0$; equivalent to p = legendre(LegendreUnitNorm(), l, 0, x).

CMB.Legendre.Plm! — Method

Plm!(P, l::Integer, m::Integer, x)

Fills the array P with the unit-normalized associated Legendre polynomial values $P_ℓ^m(x)$; equivalent to legendre!(LegendreUnitNorm(), P, l, m, x).

CMB.Legendre.Plm — Method

p = Plm(l::Integer, m::Integer, x::Number)

Computes the associated Legendre polynomials using unit normalization; equivalent to p = legendre(LegendreUnitNorm(), l, m, x).

CMB.Legendre.legendre! — Method

legendre!(norm::AbstractLegendreNorm, Λ, l::Integer, m::Integer, x)

Fills the array Λ with the Legendre polynomial values $N_ℓ^m P_ℓ^m(x)$ up to/of degree l and order m for the normalization scheme norm. Λ must be an array with between 0 and 2 more dimensions than x, with the leading dimensions having the same shape as x.

If ndims(Λ) == ndims(x), then Λ is filled with the polynomial values at x for degree l and order m.

If ndims(Λ) == ndims(x) + 1, then l is interpreted as lmax, and Λ filled with polynomial values for all degrees 0 ≤ l ≤ lmax of order m.

If ndims(Λ) == ndims(x) + 2, then l is interpreted as lmax and m as mmax, and Λ is filled with polynomial values for all degrees 0 ≤ l ≤ lmax and orders 0 ≤ m ≤ min(mmax, l).

CMB.Legendre.legendre — Method

p = legendre(norm::AbstractLegendreNorm, l::Integer, m::Integer, x::Number)
P = legendre.(norm::AbstractLegendreNorm, l, m, x)

Computes the associated Legendre polynomial $N_ℓ^m P_ℓ^m(x)$ of degree l and order m at x for the normalization scheme norm.

With broadcasting syntax, the polynomials can be computed over any iterable x. Furthermore,

If l isa Integer && m isa Integer, then the output P has the same shape as x and is filled with the polynomial values of order l and degree m.

If l isa UnitRange && m isa Integer, then l is interpreted as lmax, and the output P has one more dimension than x with the trailing dimension spanning the degrees 0 ≤ l ≤ lmax.

If l isa UnitRange && m isa UnitRange, then l is interpreted as lmax and m as mmax, and the output P has two more dimensions than x with the trailing dimensions spanning the degrees 0 ≤ l ≤ lmax and orders 0 ≤ m ≤ mmax, respectively.

Note that in second and third case, the UnitRanges must satisify first(l) == 0 and first(m) == 0.

CMB.Legendre.legendre — Method

p = legendre(norm::AbstractLegendreNorm, l::Integer, x::Number)
P = legendre.(norm::AbstractLegendreNorm, l, x)

Computes the associated Legendre polynomial assuming the order $m = 0$; equivalent to legendre(norm, l, 0, x) and legendre.(norm, l, 0, x).

CMB.Legendre.λlm! — Method

λlm!(Λ, l::Integer, m::Integer, x)

Fills the array Λ with the spherical-harmonic normalized associated Legendre polynomial values $λ_ℓ^m(x)$; equivalent to legendre!(LegendreSphereNorm(), P, l, m, x).

CMB.Legendre.λlm — Method

λ = λlm(l::Integer, m::Integer, x::Number)

Computes the associated Legendre polynomials using spherical-harmonic normalization; equivalent to λ = legendre(LegendreSphereNorm(), l, m, x).

Healpix

CMB.Healpix — Module

A module of functions implementing function which interact with the HEALPix pixel definitions. In most cases, only the RING ordering functions are being provided.

See "HEALPix: A Framework for High-Resolution Discretization and Fast Analysis of Data Distributed on the Sphere" Górski, Hivon, & Banday et al (2005) ApJ 622:759–771 arXiv: astro-ph/0409513

CMB.Healpix.UNSEEN — Constant

const UNSEEN = -1.6375e+30

Special value recognized by the libhealpix/healpy routines as an unobserved/masked pixel.

CMB.Healpix.InvalidNside — Type

InvalidNside(nside)

An invalid nside value was provided.

CMB.Healpix.InvalidPixel — Type

InvalidPixel(nside, pix)

An invalid pixel index pix was provided for the given nside.

CMB.Healpix.checkhealpix — Method

checkhealpix(nside, pix)

Throws an InvalidNside exception if nside is not a valid value or an InvalidPixel exception if pix is out of range for the given $N_\mathrm{side}$.

CMB.Healpix.checkhealpix — Method

checkhealpix(nside)

Throws an InvalidNside exception if nside is not a valid value.

CMB.Healpix.iscap — Function

iscap(nside, pix)

Test for whether the given pixel pix is in either polar cap for an nside HEALPix map.

CMB.Healpix.isequbelt — Function

isequbelt(nside, pix)

Test for whether the given pixel pix is in the equatorial belt for an nside HEALPix map.

CMB.Healpix.ishealpixok — Method

isheapixok(nside, pix)

Returns true if nside is valid and pix is in the range 0 to nside2npix(nside) - 1, otherwise false.

CMB.Healpix.ishealpixok — Method

ishealpixok(nside)

Returns true if nside is a power of two in the range 1 to 2^29, otherwise false.

CMB.Healpix.isnorth — Function

isnorth(nside, pix)

Test for whether the given pixel pix is in the northern hemisphere (including the equatorial ring) for an nside HEALPix map.

CMB.Healpix.isnorthcap — Function

isnorthcap(nside, pix)

Test for whether the given pixel pix is in the northern polar cap for an nside HEALPix map.

CMB.Healpix.isnorthequbelt — Function

isnorthequbelt(nside, pix)

Test for whether the given pixel pix is in the northern equatorial belt (including the equatorial ring) for an nside HEALPix map.

CMB.Healpix.issouth — Function

issouth(nside, pix)

Test for whether the given pixel pix is in the southern hemisphere (excluding the equatorial ring) for an nside HEALPix map.

CMB.Healpix.issouthcap — Function

issouthcap(nside, pix)

Test for whether the given pixel pix is in the southern polar cap for an nside HEALPix map.

CMB.Healpix.issouthequbelt — Function

issouthequbelt(nside, pix)

Test for whether the given pixel pix is in the southern equatorial belt (excluding the equatorial ring) for an nside HEALPix map.

CMB.Healpix.npix2nside — Method

Nisde = npix2nside(npix)

Returns the equivalent Nside corresponding to the number of pixels npix.

CMB.Healpix.nring2nside — Method

Nside = nring2nside(nring)

Returns the equivalent Nside corresponding to the number of iso-latitude rings nring.

CMB.Healpix.nside2npix — Function

Npix = nside2npix(nside)

Returns the total number of pixels Npix in an nside HEALPix map. Note that HEALPix pixel indexing is 0-based, so valid pixel values are in the range 0 to Npix - 1.

CMB.Healpix.nside2npixcap — Function

Npix = nside2npixcap(nside)

Returns the number of pixels Npix in the polar caps for the given nside HEALPix map.

CMB.Healpix.nside2npixequ — Function

Npix = nside2npixequ(nside)

Returns the number of pixels Npix in the northern hemisphere, including the equatorial ring, for the given nside HEALPix map.

CMB.Healpix.nside2nring — Function

Nring = nside2nring(nside)

Returns the number of iso-latitude rings Nring in the nside HEALPix map.

CMB.Healpix.nside2pixarea — Function

σ = nside2pixarea(nside)

Returns the surface area σ (in steradians) of each pixel in the given nside HEALPix map.

CMB.Healpix.pix2ang — Method

(θ,ϕ) = pix2ang(nside, p)

Computes the colatitude and azimuth pair (θ,ϕ) for the given pixel p.

CMB.Healpix.pix2phi — Method

ϕ = pix2phi(nside, p)

Computes the azimuth ϕ for the given pixel p. nside is the Nside resolution factor.

CMB.Healpix.pix2ring — Method

i = pix2ring(nside, p)

Computes the ring index i for the given pixel p. nside is the Nside resolution factor.

CMB.Healpix.pix2ringidx — Method

j = pix2ringidx(nside, p)

Computes the index j within the ring for the given pixel p. nside is the Nside resolution factor.

CMB.Healpix.pix2theta — Method

θ = pix2theta(nside, p)

Computes the colatitude θ for the given pixel p. nside is the Nside resolution factor.

CMB.Healpix.pix2vec — Method

r::SVector{3} = pix2vec(nside, p)

Computes the unit vector r pointing to the pixel center of the given pixel p.

CMB.Healpix.pix2z — Method

z = pix2z(nside, p)

Computes the cosine of the colatitude z for the given pixel p. nside is the Nside resolution factor.

Pixel Covariance

CMB.PixelCovariance — Module

Collection of functions which compute the pixel-pixel covariance of the CMB sky.

Based on equations given in Tegmark and de Oliveira-Costa (2001) “How to measure CMB polarization power spectra without losing information” arXiv:astro-ph/0012120v3

CMB.PixelCovariance.PixelCovarianceCache — Type

struct PixelCovarianceCache

Data structure which contains all the information and buffers required to compute the pixel-pixel covariance terms for a given pixel with respect to all other pixels.

CMB.PixelCovariance.PixelCovarianceCoeff — Type

struct PixelCovarianceCoeff{T<:Real}

Precomputed recursion relation coefficients for computing the pixel-pixel covariance.

Example

julia> PixelCovarianceCoeff{Float64}(2)
PixelCovarianceCoeff{Float64} for lmax = 2 with coefficients:
    λ: LegendreNormCoeff{LegendreUnitNorm,Float64}
    η: [0.0795775, 0.238732, 0.397887]
    α: [0.0, 0.0, 0.324874]
    β: [0.0, 0.0, 0.0331573]

Miscellaneous Utilities

CMB.Util — Module

Miscellaneous utility functions.

CMB.Util.outer — Function

Computes the outer product between a given column of a sparse matrix and a vector.

CMB.Util.outer — Method

outer(w::AbstractVector, A::SparseMatrixCSC, n::Integer)

Performs the equivalent of $\vec w \vec{a}_n^\dagger$ where $\vec a_n$ is the column A[:,n].

CMB.Util.outer — Method

outer(A::SparseMatrixCSC, n::Integer, w::AbstractVector)

Performs the equivalent of $\vec a_n \vec w^\dagger$ where $\vec a_n$ is the column A[:,n].

CMB.Util.quadprod — Function

quadprod(A, b, n, dir=:col)

Computes the quadratic product $ABA^T$ efficiently for the case where $B$ is all zero except for the nth column or row vector b, for dir = :col or dir = :row, respectively.