Public Documentation

Contents

Public Documentation

Contents

Sphere Functions

Healpix

Pixel Covariance

Sphere Functions

CMB.Sphere — Module

Collection of routines for working with coordinates on the sphere.

CMB.Sphere.bearing — Function

Calculates the bearing angle ($α$), defined as the angle between the meridian (at the first coordinate) and the great circle connecting the first coordinate to the second. Angles are measured eastward of north and will be in the range $[-π,π]$. See also bearing2.

CMB.Sphere.bearing — Method

α = bearing(θ₁, ϕ₁, θ₂, ϕ₂)

Points on the sphere are given as coordinate pairs $(θ₁,ϕ₁)$ and $(θ₂,ϕ₂)$ where $θ$ is the colatitude angle from the North Pole and $ϕ$ is the azimuthal angle, both in radians.

Examples

julia> bearing(π/2, 0.0, π/4, π/4)
0.6154797086703873

CMB.Sphere.bearing — Method

α = bearing(r₁, r₂)

Points on the sphere are given as unit vectors $r₁$ and $r₂$.

Examples

julia> bearing([1.0, 0.0, 0.0], [0.5, 0.5, sqrt(2)/2])
0.6154797086703873

CMB.Sphere.bearing2 — Function

Calculates the north/east vector components of the bearing angle (i.e. $δn = \cos(α), δe = \sin(α)$), defined as the angle between the meridian (at the first coordinate) and the great circle connecting the first coordinate to the second. See also bearing.

CMB.Sphere.bearing2 — Method

(δn, δe) = bearing2(θ₁, ϕ₁, θ₂, ϕ₂)

Points on the sphere are given as coordinate pairs $(θ₁,ϕ₁)$ and $(θ₂,ϕ₂)$ where $θ$ is the colatitude angle from the North Pole and $ϕ$ is the azimuthal angle, both in radians.

Examples

julia> bearing2(π/2, 0.0, π/4, π/4)
(0.8164965809277261, 0.5773502691896257)

CMB.Sphere.bearing2 — Method

(δn, δe) = bearing2(r₁, r₂)

Points on the sphere are given as unit vectors $r₁$ and $r₂$.

Examples

julia> bearing2([1.0, 0.0, 0.0], [0.5, 0.5, sqrt(2)/2])
(0.816496580927726, 0.5773502691896257)

CMB.Sphere.cosdistance — Function

Calculates the cosine of the inner angle ($z$) between unit vectors pointing from the center of the sphere to two points on its surface.

CMB.Sphere.cosdistance — Method

z = cosdistance(θ₁, ϕ₁, θ₂, ϕ₂)

Points on the sphere are given as coordinate pairs $(θ₁,ϕ₁)$ and $(θ₂,ϕ₂)$ where $θ$ is the colatitude angle from the North Pole and $ϕ$ is the azimuthal angle, both in radians.

Examples

julia> cosdistance(π/2, 0.0, π/4, π/4)
0.5

CMB.Sphere.cosdistance — Method

z = cosdistance(r₁, r₂)

Points on the sphere are given as unit vectors $r₁$ and $r₂$.

Examples

julia> cosdistance([1.,0.,0.], [0.5,0.5,sqrt(2)/2])
0.5

CMB.Sphere.distance — Function

Calculates the inner angle ($σ$) between unit vectors pointing from the center of the sphere to two points on its surface.

CMB.Sphere.distance — Method

σ = distance(θ₁, ϕ₁, θ₂, ϕ₂)

Points on the sphere are given as coordinate pairs $(θ₁,ϕ₁)$ and $(θ₂,ϕ₂)$ where $θ$ is the colatitude angle from the North Pole and $ϕ$ is the azimuthal angle, both in radians.

Examples

julia> distance(π/2, 0.0, π/4, π/4)
1.0471975511965979

CMB.Sphere.distance — Method

σ = distance(r₁, r₂)

Points on the sphere are given as unit vectors $r₁$ and $r₂$.

Examples

julia> distance([1.,0.,0.], [0.5,0.5,sqrt(2)/2])
1.0471975511965979

CMB.Sphere.reckon — Function

Calculates a position on the sphere a given distance (σ, in radians) and relative bearing angle (α, in radians) away from a given point (measuring the eastward-of-north orientation of the great circle connecting the source and destination points with respect to the merdian passing through the source).

CMB.Sphere.reckon — Method

(θ′, ϕ′) = reckon(θ, ϕ, σ, α)

The point on the sphere is given by the colatitude-azimuth pair (θ, ϕ), both given in radians.

Note

When r points to either the north or south pole, the meridian is defined to be θ meridian, and the bearing angle α is oriented with respect to it.

For example, moving a distance $π/2$ with no bearing goes to the equator, with the longitude dependent on the input longitude:

julia> reckon(0.0, 0.0, π/2, 0.0)
(1.5707963267948966, 3.141592653589793)

julia> reckon(0.0, π/2, π/2, 0.0)
(1.5707963267948966, 4.71238898038469)

CMB.Sphere.reckon — Method

r′ = reckon(r::AbstractVector, σ, α)

The point on the sphere is given as a unit vector r.

Note

When r points to either the north or south pole, the meridian is defined to be prime meridian and the bearing angle α is oriented with respect to it.

For example, moving a distance $π/2$ with no bearing goes to the negative $x$ axis (i.e. 0° N, 180° W):

julia> reckon([0.0, 0.0, 1.0], π/2, 0.0)
3-element StaticArrays.SArray{Tuple{3},Float64,1,3} with indices SOneTo(3):
 -1.0
  0.0
  6.123233995736766e-17

Healpix

CMB.Healpix — Module

A module of functions implementing function which interact with the HEALPix pixel definitions. In most cases, only the RING ordering functions are being provided.

See "HEALPix: A Framework for High-Resolution Discretization and Fast Analysis of Data Distributed on the Sphere" Górski, Hivon, & Banday et al (2005) ApJ 622:759–771 arXiv: astro-ph/0409513

CMB.Healpix.UNSEEN — Constant

const UNSEEN = -1.6375e+30

Special value recognized by the libhealpix/healpy routines as an unobserved/masked pixel.

CMB.Healpix.InvalidNside — Type

InvalidNside(nside)

An invalid nside value was provided.

CMB.Healpix.InvalidPixel — Type

InvalidPixel(nside, pix)

An invalid pixel index pix was provided for the given nside.

CMB.Healpix.ang2pix — Method

p = ang2pix(nside, θ, ϕ)

Computes the HEALPix pixel index p which contains the point $(θ,ϕ)$ given by the colatitude θ and azimuth ϕ, where nside is the Nside resolution factor.

CMB.Healpix.checkhealpix — Method

checkhealpix(nside, pix)

Throws an InvalidNside exception if nside is not a valid value or an InvalidPixel exception if pix is out of range for the given $N_\mathrm{side}$.

CMB.Healpix.checkhealpix — Method

checkhealpix(nside)

Throws an InvalidNside exception if nside is not a valid value.

CMB.Healpix.iscap — Function

iscap(nside, pix)

Test for whether the given pixel pix is in either polar cap for an nside HEALPix map.

CMB.Healpix.isequbelt — Function

isequbelt(nside, pix)

Test for whether the given pixel pix is in the equatorial belt for an nside HEALPix map.

CMB.Healpix.ishealpixok — Method

isheapixok(nside, pix)

Returns true if nside is valid and pix is in the range 0 to nside2npix(nside) - 1, otherwise false.

CMB.Healpix.ishealpixok — Method

ishealpixok(nside)

Returns true if nside is a power of two in the range 1 to 2^29, otherwise false.

CMB.Healpix.isnorth — Function

isnorth(nside, pix)

Test for whether the given pixel pix is in the northern hemisphere (including the equatorial ring) for an nside HEALPix map.

CMB.Healpix.isnorthcap — Function

isnorthcap(nside, pix)

Test for whether the given pixel pix is in the northern polar cap for an nside HEALPix map.

CMB.Healpix.isnorthequbelt — Function

isnorthequbelt(nside, pix)

Test for whether the given pixel pix is in the northern equatorial belt (including the equatorial ring) for an nside HEALPix map.

CMB.Healpix.issouth — Function

issouth(nside, pix)

Test for whether the given pixel pix is in the southern hemisphere (excluding the equatorial ring) for an nside HEALPix map.

CMB.Healpix.issouthcap — Function

issouthcap(nside, pix)

Test for whether the given pixel pix is in the southern polar cap for an nside HEALPix map.

CMB.Healpix.issouthequbelt — Function

issouthequbelt(nside, pix)

Test for whether the given pixel pix is in the southern equatorial belt (excluding the equatorial ring) for an nside HEALPix map.

CMB.Healpix.npix2nside — Method

Nisde = npix2nside(npix)

Returns the equivalent Nside corresponding to the number of pixels npix.

CMB.Healpix.nring2nside — Method

Nside = nring2nside(nring)

Returns the equivalent Nside corresponding to the number of iso-latitude rings nring.

CMB.Healpix.nside2npix — Function

Npix = nside2npix(nside)

Returns the total number of pixels Npix in an nside HEALPix map. Note that HEALPix pixel indexing is 0-based, so valid pixel values are in the range 0 to Npix - 1.

CMB.Healpix.nside2npixcap — Function

Npix = nside2npixcap(nside)

Returns the number of pixels Npix in the polar caps for the given nside HEALPix map.

CMB.Healpix.nside2npixequ — Function

Npix = nside2npixequ(nside)

Returns the number of pixels Npix in the northern hemisphere, including the equatorial ring, for the given nside HEALPix map.

CMB.Healpix.nside2nring — Function

Nring = nside2nring(nside)

Returns the number of iso-latitude rings Nring in the nside HEALPix map.

CMB.Healpix.nside2pixarea — Function

σ = nside2pixarea(nside)

Returns the surface area σ (in steradians) of each pixel in the given nside HEALPix map.

CMB.Healpix.pix2ang — Method

(θ,ϕ) = pix2ang(nside, p)

Computes the colatitude and azimuth pair (θ,ϕ) for the given pixel p, where nside is the Nside resolution factor.

CMB.Healpix.pix2phi — Method

ϕ = pix2phi(nside, p)

Computes the azimuth ϕ for the given pixel p, where nside is the Nside resolution factor.

CMB.Healpix.pix2ring — Method

i = pix2ring(nside, p)

Computes the ring index i for the given pixel p, where nside is the Nside resolution factor.

CMB.Healpix.pix2ringidx — Method

j = pix2ringidx(nside, p)

Computes the index j within the ring for the given pixel p, where nside is the Nside resolution factor.

CMB.Healpix.pix2theta — Method

θ = pix2theta(nside, p)

Computes the colatitude θ for the given pixel p, where nside is the Nside resolution factor.

CMB.Healpix.pix2vec — Method

r = pix2vec(nside, p)

Computes the unit vector r pointing to the pixel center of the given pixel p, where nside is the Nside resolution factor.

CMB.Healpix.pix2z — Method

z = pix2z(nside, p)

Computes the cosine of the colatitude z for the given pixel p, where nside is the Nside resolution factor.

CMB.Healpix.vec2pix — Method

p = vec2pix(nside, r)

Computes the HEALPix pixel index p which contains the point at the end of the unit vector r, where nside is the Nside resolution factor.

Pixel Covariance

The following enums/bit flags are not exported globally, but all of the named values can be imported into a scope by using the parent module (e.g. to access all of the covariance field constants, use using CMB.PixelCovariance.CovarianceFields.

CMB.PixelCovariance.PolarizationConventions.Convention — Type

@enum Convention IAUConv HealpixConv

An enumeration to specify the two types of polarization conventions used to describe Stokes Q/U coordinate systems.

CMB.PixelCovariance.CovarianceFields.Field — Type

@bitflag Field TT QT UT TQ QQ UQ TU QU UU NO_FIELD=0

A bitfield for identifying subblocks of the pixel-pixel covariance matrix. There are 9 subblocks, named as the Cartesian product of elements T, Q, and U:

TT  TQ  TU
QT  QQ  QU
UT  UQ  UU

CMB.PixelCovariance.CovarianceFields.TPol — Constant

const TPol = QT | UT | TQ | TU

An alias for the temperature-cross-polarization sub-blocks of the full covariance matrix.

CMB.PixelCovariance.CovarianceFields.Pol — Constant

const Pol  = QQ | UQ | QU | UU

An alias for the polarization-only sub-blocks of the full covariance matrix.